@@ -2,45 +2,45 @@ import 'package:flutter/material.dart';
 
 final orangeGradient = LinearGradient(colors: [
   Colors.orange,
-  Colors.orange[300],
+  Colors.orange[300]!,
 ]);
 
 final blueGradient = LinearGradient(colors: [
   Colors.blue,
-  Colors.blue[300],
+  Colors.blue[300]!,
 ]);
 
 final deepPurpleGradient = LinearGradient(colors: [
   Colors.deepPurple,
-  Colors.deepPurple[300],
+  Colors.deepPurple[300]!,
 ]);
 
 final redGradient = LinearGradient(colors: [
   Colors.red,
-  Colors.red[300],
+  Colors.red[300]!,
 ]);
 
 final greenGradient = LinearGradient(colors: [
   Colors.green,
-  Colors.green[300],
+  Colors.green[300]!,
 ]);
 
 final lightBlueGradient = LinearGradient(colors: [
   Colors.lightBlue,
-  Colors.lightBlue[300],
+  Colors.lightBlue[300]!,
 ]);
 
 final purpleGradient = LinearGradient(colors: [
   Colors.purple,
-  Colors.purple[300],
+  Colors.purple[300]!,
 ]);
 
 final lightGreenGradient = LinearGradient(colors: [
   Colors.lightGreen,
-  Colors.lightGreen[300],
+  Colors.lightGreen[300]!,
 ]);
 
 final blueGreyGradient = LinearGradient(colors: [
   Colors.blueGrey,
-  Colors.blueGrey[300],
+  Colors.blueGrey[300]!,
 ]);
